temporary branch, added basis for handling boundary conditions
improved a few things with multivariate PDF generators initialization

... but forgot to commit changes about proper rotations matrices in the FB algo...

+++ GeneralUtils/LinearTransformUtils/3DTransformUtils.h
@@ -51,7 +51,7 @@ typedef enum {
 /** Convention for euler angles is heading, attitude, bank 
 * returns a 3*3 rotation matrix
 */
-vnl_matrix<double> Get3DRotationMatrixFromEulerAngles(float h, float a, float b);
+vnl_matrix<double> Get3DRotationMatrixFromEulerAngles(double h, double a, double b);
 
 /** Convention for euler angles is heading, attitude, bank 
 * returns a 3*3 rotation matrix
@@ -77,16 +77,16 @@ vnl_matrix<double> Rotate3DTransformAroundZ(const vnl_matrix<double> &R, double 
 void Rotate3DTransformAroundZ_InPlace(vnl_matrix<double> &R, double angle);
 
 
-/** 3D rotation matrix to Quaternion */
+/** 3D rotation matrix to Quaternion (w, x, y, z) */
 vnl_vector<double> GetQuaternionFrom3DRotationMatrix(const vnl_matrix<double> &R);
-/** Quaternion to 3D rotation matrix (outputs a 3*3 matrices) */
+/** Quaternion (w, x, y, z)  to 3D rotation matrix (outputs a 3*3 matrices) */
 vnl_matrix<double> Get3DRotationMatrixFromQuaternion_33(const vnl_vector<double> &Q);
-/** Quaternion to 3D rotation matrix (outputs a 4*4 matrices) */
+/** Quaternion (w, x, y, z)  to 3D rotation matrix (outputs a 4*4 matrices) */
 vnl_matrix<double> Get3DRotationMatrixFromQuaternion_44(const vnl_vector<double> &Q);
 
-/** Get vector and angle from quaternion */
+/** Get vector and angle from quaternion (w, x, y, z)  */
 void GetVectorAndAngleFromQuaternion(vnl_vector<double> &v, double &a, const vnl_vector<double> &Q);
-/** Get quaternion from vector and angle */
+/** Get quaternion (w, x, y, z) from vector and angle */
 void GetQuaternionFromVectorAndAngle(const vnl_vector<double> &v, double a, vnl_vector<double> &Q);
 
 } // namespace psciob
